Move UTXO_WALLET_STATE to mempool
