+++ scripts/charting.js
@@ -11,7 +11,7 @@ import { cChainParams, COIN } from './chain_params';
 import { doms, mempool } from './global';
 import { Database } from './database.js';
 import { translation } from './i18n';
-import { UTXO_WALLET_STATE } from './wallet';
+import { UTXO_WALLET_STATE } from './mempool';
 import { COutpoint } from './mempool';
 
 Chart.register(
